kfa-class {kernlab}	R Documentation

Class "kfa"

Description

The class of the object returned by the Kernel Feature Analysis kfa function

Objects from the Class

Objects can be created by calls of the form new("kfa", ...) or by calling the kfa method. The objects contain the features along with the alpha values.

Slots

alpha:

Object of class "matrix" containing the alpha values

alphaindex:

Object of class "vector" containing the indexes of the selected feature

kernelf:

Object of class "kfunction" containing the kernel function used

xmatrix:

Object of class "matrix" containing the selected features

kcall:

Object of class "call" containing the kfa function call

terms:

Object of class "ANY" containing the formula terms

Methods

alpha

signature(object = "kfa"): returns the alpha values

alphaindex

signature(object = "kfa"): returns the index of the selected features

kcall

signature(object = "kfa"): returns the function call

kernelf

signature(object = "kfa"): returns the kernel function used

predict

signature(object = "kfa"): used to embed more data points to the feature base

xmatrix

signature(object = "kfa"): returns the selected features.

Author(s)

Alexandros Karatzoglou
alexandros.karatzoglou@ci.tuwien.ac.at

See Also

kfa, kpca-class

Examples

data(promotergene)
f <- kfa(~.,data=promotergene)

[Package kernlab version 0.9-27 Index]