lssvm-class {kernlab}	R Documentation

Class "lssvm"

Description

The Gaussian Processes object

Objects from the Class

Objects can be created by calls of the form new("lssvm", ...). or by calling the lssvm function

Slots

kernelf:

Object of class "kfunction" contains the kernel function used

kpar:

Object of class "list" contains the kernel parameter used

param:

Object of class "list" contains the regularization parameter used.

kcall:

Object of class "call" contains the used function call

type:

Object of class "character" contains type of problem

coef:

Object of class "ANY" contains the model parameter

terms:

Object of class "ANY" contains the terms representation of the symbolic model used (when using a formula)

xmatrix:

Object of class "matrix" containing the data matrix used

ymatrix:

Object of class "output" containing the response matrix

fitted:

Object of class "output" containing the fitted values

b:

Object of class "numeric" containing the offset

lev:

Object of class "vector" containing the levels of the response (in case of classification)

scaling:

Object of class "ANY" containing the scaling information performed on the data

nclass:

Object of class "numeric" containing the number of classes (in case of classification)

alpha:

Object of class "listI" containing the computes alpha values

alphaindex

Object of class "list" containing the indexes for the alphas in various classes (in multi-class problems).

error:

Object of class "numeric" containing the training error

cross:

Object of class "numeric" containing the cross validation error

n.action:

Object of class "ANY" containing the action performed in NA

nSV:

Object of class "numeric" containing the number of model parameters

Methods

alpha

signature(object = "lssvm"): returns the alpha vector

cross

signature(object = "lssvm"): returns the cross validation error

error

signature(object = "lssvm"): returns the training error

fitted

signature(object = "vm"): returns the fitted values

kcall

signature(object = "lssvm"): returns the call performed

kernelf

signature(object = "lssvm"): returns the kernel function used

kpar

signature(object = "lssvm"): returns the kernel parameter used

param

signature(object = "lssvm"): returns the regularization parameter used

lev

signature(object = "lssvm"): returns the response levels (in classification)

type

signature(object = "lssvm"): returns the type of problem

scaling

signature(object = "ksvm"): returns the scaling values

xmatrix

signature(object = "lssvm"): returns the data matrix used

ymatrix

signature(object = "lssvm"): returns the response matrix used

Author(s)

Alexandros Karatzoglou
alexandros.karatzoglou@ci.tuwien.ac.at

See Also

lssvm, ksvm-class

Examples

# train model
data(iris)
test <- lssvm(Species~.,data=iris,var=2)
test
alpha(test)
error(test)
lev(test)

[Package kernlab version 0.9-27 Index]