predict.ksvm {kernlab}	R Documentation

predict method for support vector object

Description

Prediction of test data using support vector machines

Usage

## S4 method for signature 'ksvm'
predict(object, newdata, type = "response", coupler = "minpair")

Arguments

object
an S4 object of class ksvm created by the ksvm function

newdata
a data frame or matrix containing new data

type
one of response, probabilities ,votes, decision indicating the type of output: predicted values, matrix of class probabilities, matrix of vote counts, or matrix of decision values.

coupler
Coupling method used in the multiclass case, can be one of minpair or pkpd (see reference for more details).

Value

If type(object) is C-svc, nu-svc, C-bsvm or spoc-svc the vector returned depends on the argument type:

response
predicted classes (the classes with majority vote).

probabilities
matrix of class probabilities (one column for each class and one row for each input).

votes
matrix of vote counts (one column for each class and one row for each new input)

If type(object) is eps-svr, eps-bsvr or nu-svr a vector of predicted values is returned. If type(object) is one-classification a vector of logical values is returned.

Author(s)

Alexandros Karatzoglou
alexandros.karatzoglou@ci.tuwien.ac.at

References

T.F. Wu, C.J. Lin, R.C. Weng.
Probability estimates for Multi-class Classification by Pairwise Coupling
http://www.csie.ntu.edu.tw/~cjlin/papers/svmprob/svmprob.pdf

H.T. Lin, C.J. Lin, R.C. Weng
A note on Platt's probabilistic outputs for support vector machines
http://www.csie.ntu.edu.tw/~cjlin/papers/plattprob.pdf

Examples

## example using the promotergene data set
data(promotergene)

## create test and training set
ind <- sample(1:dim(promotergene)[1],20)
genetrain <- promotergene[-ind, ]
genetest <- promotergene[ind, ]

## train a support vector machine
gene <- ksvm(Class~.,data=genetrain,kernel="rbfdot",
             kpar=list(sigma=0.015),C=70,cross=4,prob.model=TRUE)
gene

## predict gene type probabilities on the test set
genetype <- predict(gene,genetest,type="probabilities")
genetype

[Package kernlab version 0.9-27 Index]