ranking-class {kernlab}	R Documentation

Class "ranking"

Description

Object of the class "ranking" are created from the ranking function and extend the class matrix

Objects from the Class

Objects can be created by calls of the form new("ranking", ...).

Slots

.Data:

Object of class "matrix" containing the data ranking and scores

convergence:

Object of class "matrix" containing the convergence matrix

edgegraph:

Object of class "matrix" containing the edgegraph

Extends

Class "matrix", directly.

Methods

show

signature(object = "ranking"): displays the ranking score matrix

Author(s)

Alexandros Karatzoglou
alexandros.karatzoglou@ci.tuwien.ac.at

See Also

ranking

Examples

data(spirals)

## create data set to be ranked
ran<-spirals[rowSums(abs(spirals)<0.55)==2,]

## rank points according to "relevance" to point 54 (up left)
ranked<-ranking(ran,54,kernel="rbfdot",
                kpar=list(sigma=100),edgegraph=TRUE)

ranked
edgegraph(ranked)[1:10,1:10]

[Package kernlab version 0.9-27 Index]