reuters {kernlab}	R Documentation

Reuters Text Data

Description

A small sample from the Reuters news data set.

Usage

data(reuters)

Format

A list of 40 text documents along with the labels. reuters contains the text documents and rlabels the labels in a vector.

Details

This dataset contains a list of 40 text documents along with the labels. The data consist out of 20 documents from the acq category and 20 documents from the crude category. The labels are stored in rlabels

Source

Reuters

[Package kernlab version 0.9-27 Index]