rvm-class {kernlab}	R Documentation

Class "rvm"

Description

Relevance Vector Machine Class

Objects from the Class

Objects can be created by calls of the form new("rvm", ...). or by calling the rvm function.

Slots

tol:

Object of class "numeric" contains tolerance of termination criteria used.

kernelf:

Object of class "kfunction" contains the kernel function used

kpar:

Object of class "list" contains the hyperparameter used

kcall:

Object of class "call" contains the function call

type:

Object of class "character" contains type of problem

terms:

Object of class "ANY" containing the terms representation of the symbolic model used (when using a formula interface)

xmatrix:

Object of class "matrix" contains the data matrix used during computation

ymatrix:

Object of class "output" contains the response matrix

fitted:

Object of class "output" with the fitted values, (predict on training set).

lev:

Object of class "vector" contains the levels of the response (in classification)

nclass:

Object of class "numeric" contains the number of classes (in classification)

alpha:

Object of class "listI" containing the the resulting alpha vector

coef:

Object of class "ANY" containing the the resulting model parameters

nvar:

Object of class "numeric" containing the calculated variance (in case of regression)

mlike:

Object of class "numeric" containing the computed maximum likelihood

RVindex:

Object of class "vector" containing the indexes of the resulting relevance vectors

nRV:

Object of class "numeric" containing the number of relevance vectors

cross:

Object of class "numeric" containing the resulting cross validation error

error:

Object of class "numeric" containing the training error

n.action:

Object of class "ANY" containing the action performed on NA

Methods

RVindex

signature(object = "rvm"): returns the index of the relevance vectors

alpha

signature(object = "rvm"): returns the resulting alpha vector

cross

signature(object = "rvm"): returns the resulting cross validation error

error

signature(object = "rvm"): returns the training error

fitted

signature(object = "vm"): returns the fitted values

kcall

signature(object = "rvm"): returns the function call

kernelf

signature(object = "rvm"): returns the used kernel function

kpar

signature(object = "rvm"): returns the parameters of the kernel function

lev

signature(object = "rvm"): returns the levels of the response (in classification)

mlike

signature(object = "rvm"): returns the estimated maximum likelihood

nvar

signature(object = "rvm"): returns the calculated variance (in regression)

type

signature(object = "rvm"): returns the type of problem

xmatrix

signature(object = "rvm"): returns the data matrix used during computation

ymatrix

signature(object = "rvm"): returns the used response

Author(s)

Alexandros Karatzoglou
alexandros.karatzoglou@ci.tuwien.ac.at

See Also

rvm, ksvm-class

Examples

# create data
x <- seq(-20,20,0.1)
y <- sin(x)/x + rnorm(401,sd=0.05)

# train relevance vector machine
foo <- rvm(x, y)
foo

alpha(foo)
RVindex(foo)
fitted(foo)
kernelf(foo)
nvar(foo)

## show slots
slotNames(foo)

[Package kernlab version 0.9-27 Index]