specc-class {kernlab}	R Documentation

Class "specc"

Description

The Spectral Clustering Class

Objects from the Class

Objects can be created by calls of the form new("specc", ...). or by calling the function specc.

Slots

.Data:

Object of class "vector" containing the cluster assignments

centers:

Object of class "matrix" containing the cluster centers

size:

Object of class "vector" containing the number of points in each cluster

withinss:

Object of class "vector" containing the within-cluster sum of squares for each cluster

kernelf

Object of class kernel containing the used kernel function.

Methods

centers

signature(object = "specc"): returns the cluster centers

withinss

signature(object = "specc"): returns the within-cluster sum of squares for each cluster

size

signature(object = "specc"): returns the number of points in each cluster

Author(s)

Alexandros Karatzoglou
alexandros.karatzoglou@ci.tuwien.ac.at

See Also

specc, kpca-class

Examples

## Cluster the spirals data set.
data(spirals)

sc <- specc(spirals, centers=2)

centers(sc)
size(sc)

[Package kernlab version 0.9-27 Index]