vm-class {kernlab}	R Documentation

Class "vm"

Description

An S4 VIRTUAL class used as a base for the various vector machine classes in kernlab

Objects from the Class

Objects from the class cannot be created directly but only contained in other classes.

Slots

alpha:

Object of class "listI" containing the resulting alpha vector (list in case of multiclass classification) (support vectors)

type:

Object of class "character" containing the vector machine type e.g., ("C-svc", "nu-svc", "C-bsvc", "spoc-svc", "one-svc", "eps-svr", "nu-svr", "eps-bsvr")

kernelf:

Object of class "function" containing the kernel function

kpar:

Object of class "list" containing the kernel function parameters (hyperparameters)

kcall:

Object of class "call" containing the function call

terms:

Object of class "ANY" containing the terms representation of the symbolic model used (when using a formula)

xmatrix:

Object of class "input" the data matrix used during computations (support vectors) (possibly scaled and without NA)

ymatrix:

Object of class "output" the response matrix/vector

fitted:

Object of class "output" with the fitted values, predictions using the training set.

lev:

Object of class "vector" with the levels of the response (in the case of classification)

nclass:

Object of class "numeric" containing the number of classes (in the case of classification)

error:

Object of class "vector" containing the training error

cross:

Object of class "vector" containing the cross-validation error

n.action:

Object of class "ANY" containing the action performed for NA

Methods

alpha

signature(object = "vm"): returns the complete alpha vector (wit zero values)

cross

signature(object = "vm"): returns the cross-validation error

error

signature(object = "vm"): returns the training error

fitted

signature(object = "vm"): returns the fitted values (predict on training set)

kernelf

signature(object = "vm"): returns the kernel function

kpar

signature(object = "vm"): returns the kernel parameters (hyperparameters)

lev

signature(object = "vm"): returns the levels in case of classification

kcall

signature(object="vm"): returns the function call

type

signature(object = "vm"): returns the problem type

xmatrix

signature(object = "vm"): returns the data matrix used(support vectors)

ymatrix

signature(object = "vm"): returns the response vector

Author(s)

Alexandros Karatzoglou
alexandros.karatzolgou@ci.tuwien.ac.at

See Also

ksvm-class, rvm-class, gausspr-class

[Package kernlab version 0.9-27 Index]